Move controllers package back into the model
I hate this, but it avoids import cycles. Hopefully it can be refactored in the future.
